# translation-layer/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "translation-layer",
        "description": "",
        "contact": {
            "name": "Space and Time Labs"
        },
        "license": {
            "name": "MIT-0",
            "identifier": "MIT-0"
        },
        "version": "0.1.0"
    },
    "paths": {
        "/api/mainnet/add_smartcontract": {
            "post": {
                "tags": [
                    "add-smartcontract"
                ],
                "summary": "Submits a transaction to add a smart contract to the indexing system.",
                "description": "This endpoint takes a smart contract configuration, converts it into the appropriate runtime format,\nand submits it as an extrinsic to the blockchain.\n\n# Request Body\n- `source` (String): The source blockchain network.\n- `address` (String): The contract's address on the source chain.\n- `abi` (Option<String>): The optional Application Binary Interface (ABI) of the contract.\n- `starting_block` (Option<u64>): The block number from which to start indexing.\n- `target_schema` (Option<String>): The schema or DDL statement for contract data.\n- `contract_name` (Option<String>): A human-readable name for the contract.\n- `events` (Vec<EventMapping>): List of event mappings (origin event → target table).\n- `ddl_statement` (Option<String>): The DDL statement corresponding to the smart contract.\n\n# Responses\n- **200 OK**: Smart contract successfully added.\n- **400 BAD REQUEST**: Invalid request parameters.\n- **500 INTERNAL SERVER ERROR**: Transaction submission failed.\n\n# Example Usage\n```sh\ncurl -X POST \"http://127.0.0.1:3000/add_smartcontract\" -H \"Content-Type: application/json\" -d '{...}'\n```",
                "operationId": "add_smartcontract",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddContractRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Smart contract added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/mainnet/get_smartcontract": {
            "get": {
                "tags": [
                    "/get-smartcontract"
                ],
                "summary": "Retrieves the details of a specific smart contract.",
                "description": "This endpoint queries the blockchain storage for a smart contract by its source and address.\n\n# Query Parameters\n- `source` (String): The blockchain network of the smart contract.\n- `address` (String): The address of the smart contract.\n\n# Responses\n- **200 OK**: Successfully retrieved the smart contract details.\n- **404 NOT FOUND**: Smart contract not found.\n- **500 INTERNAL SERVER ERROR**: Error accessing blockchain storage.\n\n# Example Usage\n```sh\ncurl -X GET \"http://127.0.0.1:3000/get_smartcontract?source=Ethereum&address=0x123...\"\n```",
                "operationId": "get_smartcontract",
                "parameters": [
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Source of the smart contract",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address of the smart contract",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved contract",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetContractResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contract not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/mainnet/get_smartcontracts": {
            "get": {
                "tags": [
                    "get-smartcontracts"
                ],
                "summary": "Retrieves a list of all smart contracts indexed for a given blockchain source.",
                "description": "This endpoint iterates over the blockchain storage to fetch all smart contracts stored for a specified source.\n\n# Query Parameters\n- `source` (String): The blockchain network whose smart contracts should be retrieved.\n\n# Responses\n- **200 OK**: Successfully retrieved the list of contracts.\n- **404 NOT FOUND**: No contracts found for the given source.\n- **500 INTERNAL SERVER ERROR**: Error accessing blockchain storage.\n\n# Example Usage\n```sh\ncurl -X GET \"http://127.0.0.1:3000/get_smartcontracts?source=Ethereum\"\n```",
                "operationId": "get_smartcontracts",
                "parameters": [
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Source of the smart contracts",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved contracts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetContractsResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No contracts found for the given source",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/mainnet/remove_smartcontract": {
            "post": {
                "tags": [
                    "remove-smartcontract"
                ],
                "summary": "Submits a transaction to remove a smart contract from the indexing system.",
                "description": "This endpoint allows for the removal of an indexed smart contract from the blockchain storage.\n\n# Request Body\n- `source` (String): The source blockchain network.\n- `address` (String): The contract's address on the source chain.\n\n# Responses\n- **200 OK**: Smart contract successfully removed.\n- **400 BAD REQUEST**: Invalid request parameters.\n- **500 INTERNAL SERVER ERROR**: Transaction submission failed.\n\n# Example Usage\n```sh\ncurl -X POST \"http://127.0.0.1:3000/remove_smartcontract\" -H \"Content-Type: application/json\" -d '{...}'\n```",
                "operationId": "remove_smartcontract",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RemoveContractRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Smart contract removed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/testnet/add_smartcontract": {
            "post": {
                "tags": [
                    "add-smartcontract"
                ],
                "summary": "Submits a transaction to add a smart contract to the indexing system.",
                "description": "This endpoint takes a smart contract configuration, converts it into the appropriate runtime format,\nand submits it as an extrinsic to the blockchain.\n\n# Request Body\n- `source` (String): The source blockchain network.\n- `address` (String): The contract's address on the source chain.\n- `abi` (Option<String>): The optional Application Binary Interface (ABI) of the contract.\n- `starting_block` (Option<u64>): The block number from which to start indexing.\n- `target_schema` (Option<String>): The schema or DDL statement for contract data.\n- `contract_name` (Option<String>): A human-readable name for the contract.\n- `events` (Vec<EventMapping>): List of event mappings (origin event → target table).\n- `ddl_statement` (Option<String>): The DDL statement corresponding to the smart contract.\n\n# Responses\n- **200 OK**: Smart contract successfully added.\n- **400 BAD REQUEST**: Invalid request parameters.\n- **500 INTERNAL SERVER ERROR**: Transaction submission failed.\n\n# Example Usage\n```sh\ncurl -X POST \"http://127.0.0.1:3000/add_smartcontract\" -H \"Content-Type: application/json\" -d '{...}'\n```",
                "operationId": "add_smartcontract",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddContractRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Smart contract added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/testnet/get_smartcontract": {
            "get": {
                "tags": [
                    "/get-smartcontract"
                ],
                "summary": "Retrieves the details of a specific smart contract.",
                "description": "This endpoint queries the blockchain storage for a smart contract by its source and address.\n\n# Query Parameters\n- `source` (String): The blockchain network of the smart contract.\n- `address` (String): The address of the smart contract.\n\n# Responses\n- **200 OK**: Successfully retrieved the smart contract details.\n- **404 NOT FOUND**: Smart contract not found.\n- **500 INTERNAL SERVER ERROR**: Error accessing blockchain storage.\n\n# Example Usage\n```sh\ncurl -X GET \"http://127.0.0.1:3000/get_smartcontract?source=Ethereum&address=0x123...\"\n```",
                "operationId": "get_smartcontract",
                "parameters": [
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Source of the smart contract",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address of the smart contract",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved contract",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetContractResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contract not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/testnet/get_smartcontracts": {
            "get": {
                "tags": [
                    "get-smartcontracts"
                ],
                "summary": "Retrieves a list of all smart contracts indexed for a given blockchain source.",
                "description": "This endpoint iterates over the blockchain storage to fetch all smart contracts stored for a specified source.\n\n# Query Parameters\n- `source` (String): The blockchain network whose smart contracts should be retrieved.\n\n# Responses\n- **200 OK**: Successfully retrieved the list of contracts.\n- **404 NOT FOUND**: No contracts found for the given source.\n- **500 INTERNAL SERVER ERROR**: Error accessing blockchain storage.\n\n# Example Usage\n```sh\ncurl -X GET \"http://127.0.0.1:3000/get_smartcontracts?source=Ethereum\"\n```",
                "operationId": "get_smartcontracts",
                "parameters": [
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Source of the smart contracts",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved contracts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetContractsResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No contracts found for the given source",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponse": {
                "type": "object",
                "description": "Generic API response structure.",
                "required": [
                    "success"
                ],
                "properties": {
                    "errMsg": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "An optional error message in case of failure."
                    },
                    "success": {
                        "type": "boolean",
                        "description": "Indicates whether the request was successful."
                    },
                    "txHash": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The transaction hash associated with the request, if applicable."
                    }
                }
            },
...
            "GetContractResponse": {
                "type": "object",
                "description": "Represents the response format when retrieving a contract's details.",
                "required": [
                    "success"
                ],
                "properties": {
                    "contract": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ApiContract",
                                "description": "The retrieved contract details, if found."
                            }
                        ]
                    },
                    "errMsg": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "An optional error message in case of failure."
                    },
                    "success": {
                        "type": "boolean",
                        "description": "Indicates whether the request was successful."
                    }
                }
            },
...
            "RemoveContractRequest": {
                "type": "object",
                "description": "Represents a request to remove a smart contract from the indexing system.",
                "required": [
                    "source",
                    "address"
                ],
                "properties": {
                    "address": {
                        "type": "string",
                        "description": "The contract's address on the source chain."
                    },
                    "source": {
                        "type": "string",
                        "description": "The source blockchain network."
                    }
                }
            },
...
        }
    }
}